remove pyc from index

+++ cugl/external/libdatachannel/deps/libsrtp/crypto/math/datatypes.c
@@ -203,7 +203,7 @@ int bitvector_alloc(bitvector_t *v, unsigned long length)
             return -1;
         }
     }
-    v->length = length;
+    v->length = (unsigned int)length;
 
     /* initialize bitvector to zero */
     bitvector_set_to_zero(v);
